% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EGA.R
\name{EGA}
\alias{EGA}
\title{Apply the Exploratory Graph Analysis technique}
\usage{
EGA(data, plot.EGA = TRUE)
}
\arguments{
\item{data}{A dataframe with the variables to be used in the analysis.}

\item{plot.EGA}{Logical. If true, returns a plot of the network of partial correlations estimated via graphical lasso and its estimated dimensions.}
}
\description{
\code{EGA} Estimates the number of dimensions of a given dataset/instrument using graphical lasso and a random walk algorithm. The glasso regularization parameter
is set via EBIC.
}
\examples{
ega.wmt <- EGA(data = wmt2[,7:24], plot.EGA = TRUE)
ega.intel <- EGA(data = intelligenceBattery[,8:66])

\dontrun{
EGA(a)
}
}
\author{
Hudson F. Golino <hfgolino at gmail.com>
}
\seealso{
\code{\link{bootEGA}} to investigate the stability of EGA's estimation via bootstrap and \code{\link{CFA}} to
verify the fit of the structure suggested by EGA using confirmatory factor analysis.
}

