% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ega.R
\docType{data}
\name{glucose_data}
\alias{glucose_data}
\title{5072 paired reference and test glucose values.}
\format{A data frame with 5072 rows and 2 variables:
\describe{
  \item{ref}{Reference method glucose value, in mg/dL}
  \item{test}{Test method glucose value, in mg/dL}
}}
\source{
The data is from a modified clinical dataset.
}
\usage{
glucose_data
}
\description{
A dataset containing 5072 paired reference method and test method
glucose values (in mg/dL).
}
\keyword{datasets}

